requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
altair
annotated-types
anyio
attrs
blinker
cachetools
certifi
charset-normalizer
click
colorama
dataclasses-json
et_xmlfile
faiss-cpu
Faker
fastapi
filetype
frozenlist
gitdb
GitPython
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
googleapis-common-protos
google-genai
greenlet
grpcio
grpcio-status
h11
httpcore
httplib2
httpx
httpx-sse
idna
Jinja2
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
langsmith
Markdown
MarkupSafe
marshmallow
multidict
mypy_extensions
narwhals
numpy
openpyxl
orjson
packaging
pandas
pillow
propcache
proto-plus
protobuf
pyarrow
pyasn1
pyasn1_modules
pydantic
pydantic-settings
pydantic_core
pydeck
pyparsing
PyPDF2
python-dateutil
python-dotenv
python-multipart
pytz
PyYAML
referencing
requests
requests-toolbelt
rpds-py
rsa
six
smmap
sniffio
SQLAlchemy
starlette
streamlit
tenacity
toml
tornado
tqdm
typing-inspect
typing-inspection
typing_extensions
tzdata
uritemplate
urllib3
uvicorn
watchdog
yarl
zstandard
